AIのScript作成
PlayerMove.cs
AI関連を分離

PlayerAI.cs:新規作成
 プレイヤーAIの動き
