+++ TeamProjectProto/Assets/PlayerAI.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 84a314f189138144186228e5fbedeb39
+timeCreated: 1532665543
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
